+++ frontends/desktop/app/Http/Controllers/AuthController.php
@@ -105,6 +105,7 @@ class AuthController extends DesktopController
             ]);
         }
 
+        $request->session()->invalidate();
         $request->session()->regenerateToken();
 
         $message = $forgetRememberedState